Organize groups of apps by packages
