+++ docs/reference/apps/crops-split.md
@@ -1,1 +1,1 @@
-::: pyodi.apps.crops_split
+::: pyodi.apps.crops.crops_split
