+++ docs/reference/apps/train-config-generation.md
@@ -1,1 +1,1 @@
-::: pyodi.apps.train_config_generation
+::: pyodi.apps.train_config.train_config_generation
