+++ pyodi/apps/coco/__init__.py
@@ -1,9 +1,7 @@
 import typer
 
-from pyodi.apps.coco_merge import coco_merge
-from pyodi.apps.coco_split import property_split, random_split
-
-# TODO Split entrypoint and logic/command modules in packages? (issue #138)
+from pyodi.apps.coco.coco_merge import coco_merge
+from pyodi.apps.coco.coco_split import property_split, random_split
 
 coco_app = typer.Typer()
 coco_app.command("merge")(coco_merge)
