+++ pyodi/apps/crops/crops_merge.py
